Make size_of, align_of, and element_offset functions return u64 instead of uint in trans::machine (per #5172)

+++ src/librustc/middle/trans/intrinsic.rs
@@ -247,7 +247,7 @@ pub fn trans_intrinsic(ccx: @CrateContext,
         "size_of" => {
             let tp_ty = substs.tys[0];
             let lltp_ty = type_of::type_of(ccx, tp_ty);
-            Ret(bcx, C_uint(ccx, machine::llsize_of_real(ccx, lltp_ty)));
+            Ret(bcx, C_uint(ccx, machine::llsize_of_real(ccx, lltp_ty) as uint));
         }
         "move_val_init" => {
             // Create a datum reflecting the value being moved.
@@ -266,12 +266,12 @@ pub fn trans_intrinsic(ccx: @CrateContext,
         "min_align_of" => {
             let tp_ty = substs.tys[0];
             let lltp_ty = type_of::type_of(ccx, tp_ty);
-            Ret(bcx, C_uint(ccx, machine::llalign_of_min(ccx, lltp_ty)));
+            Ret(bcx, C_uint(ccx, machine::llalign_of_min(ccx, lltp_ty) as uint));
         }
         "pref_align_of"=> {
             let tp_ty = substs.tys[0];
             let lltp_ty = type_of::type_of(ccx, tp_ty);
-            Ret(bcx, C_uint(ccx, machine::llalign_of_pref(ccx, lltp_ty)));
+            Ret(bcx, C_uint(ccx, machine::llalign_of_pref(ccx, lltp_ty) as uint));
         }
         "get_tydesc" => {
             let tp_ty = substs.tys[0];
@@ -337,7 +337,7 @@ pub fn trans_intrinsic(ccx: @CrateContext,
                         _ => fail!("transmute has non-expr arg"),
                     }
                 };
-                let pluralize = |n| if 1u == n { "" } else { "s" };
+                let pluralize = |n| if 1 == n { "" } else { "s" };
                 ccx.sess.span_fatal(sp,
                                     format!("transmute called on types with \
                                           different sizes: {} ({} bit{}) to \
